fix(reality-server): widen portal_saved_searches.match_count to bigint
Follow-up to PR #2812. RealityPortalRepository::mark_saved_search_matched
advances the alert watermark with `match_count = match_count + N`, and
`portal_saved_searches.match_count` was `int4`. Since PR #2812 made the
enqueue + advance atomic (correct), a failed advance now rolls the whole
run back — which exposes a latent overflow edge: a long-lived saved
search accumulating `match_count` near i32::MAX would have every
subsequent advance abort on `integer` overflow, wedging the search and
silently stopping alert delivery forever.

Widen the column to `bigint` (migration 00232), lift the Rust model /
repo types from i32 -> i64, and lock the behavior in with a new
regression test (`watermark_advances_past_i32_max_boundary`) that primes
match_count past i32::MAX and asserts the watermark still moves. The
pre-existing rollback-on-overflow guard is preserved by climbing to
i64::MAX to force the bigint overflow — same contract, new ceiling.

Closes #2814

+++ backend/crates/db/src/models/reality_portal.rs
@@ -114,7 +114,11 @@ pub struct PortalSavedSearch {
     pub alerts_enabled: bool,
     pub alert_frequency: String,
     pub last_matched_at: Option<DateTime<Utc>>,
-    pub match_count: i32,
+    /// Running total of listings matched by this saved search across every
+    /// alert run. Widened to `bigint` (i64) in migration 00232 so a long-lived,
+    /// high-traffic search can never wedge itself on integer overflow — see
+    /// #2814 and the `saved_search_watermark_advance_tests` suite.
+    pub match_count: i64,
     pub last_alert_sent_at: Option<DateTime<Utc>>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
